Use Interpolator for envelope DAC
